@@ -10,9 +10,9 @@
     "@nomiclabs/hardhat-waffle": "^2.0.3",
     "@typechain/ethers-v5": "^10.1.0",
     "@typechain/hardhat": "^6.1.2",
-    "@types/chai": "^4.2.0",
-    "@types/mocha": "^9.1.0",
-    "@types/node": ">=12.0.0",
+    "@types/chai": "^4.3.3",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.7.14",
     "chai": "^4.3.6",
     "ethereum-waffle": "^3.4.4",
     "ethers": "^5.7.0",
@@ -20,9 +20,9 @@
     "hardhat-abi-exporter": "^2.10.0",
     "hardhat-gas-reporter": "^1.0.8",
     "solidity-coverage": "^0.7.21",
-    "ts-node": ">=8.0.0",
+    "ts-node": "^10.9.1",
     "typechain": "^8.1.0",
-    "typescript": ">=4.5.0"
+    "typescript": "^4.8.2"
   },
   "dependencies": {
     "@openzeppelin/contracts": "^4.7.3",
